Use double for kernel energy calculations
Agent-Logs-Url: https://github.com/chrxh/alien/sessions/d5e59795-6190-42dc-a11e-26334d154e3c

+++ source/EngineImpl/DescConverterService.cpp
@@ -1219,7 +1219,7 @@ void DescConverterService::convertCreatureToTO(
 
 namespace
 {
-    void checkAndCorrectInvalidEnergy(float& energy)
+    void checkAndCorrectInvalidEnergy(double& energy)
     {
         if (std::isnan(energy) || energy < 0 || energy > 1e12) {
             energy = 0;
